list movies

+++ entity/src/main/java/com/example/entity/ResultsMovies.kt
@@ -10,12 +10,12 @@ data class ResultsMovies(
     var original_language: String? = "",
     var original_title: String? = "",
     var overview: String? = "",
-    var popularity: Int? = 0,
+    var popularity: Double? = 0.0,
     var poster_path: String? = "",
     var release_date: String? = "",
     var title: String? = "",
     var video: Boolean? = false,
-    var vote_average: Int? = 0,
+    var vote_average: Double? = 0.0,
     var vote_count: Int? = 0
 ): Serializable {
